requirements.txt
# Dependencies for iam_risk_assessment.py
boto3>=1.26.0
botocore>=1.29.0
pyyaml>=6.0

# Testing dependencies
hypothesis>=6.0.0
pytest>=7.0.0